pyproject.toml
[project]
name = "peso-tracker"
version = "0.1.0"
description = "ETL Cotizacion del Peso Uruguayo"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.6",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.36",
    "sqlmodel>=0.0.22",
    "supabase>=2.10.0",
    "uvicorn>=0.34.0",
    "validators>=0.34.0",
]
